Move Voiceover & CommuictyContribution validators into suggestion_validators
